+++ lisp/eab-org-src-babel.el
@@ -30,8 +30,7 @@
 
 (add-to-list 'org-babel-default-header-args:dot '(:cmdline . "-Tsvg"))
 
-;; TODO it's danger
-(setq org-confirm-babel-evaluate nil)
+(setq org-confirm-babel-evaluate 't)
 
 (setq org-edit-src-content-indentation 0)
 (setq org-src-preserve-indentation 't)